+++ templates/arc/remix.config.js
@@ -1,12 +1,14 @@
 /** @type {import('@remix-run/dev').AppConfig} */
 module.exports = {
-  future: {
-    v2_routeConvention: true,
-  },
   ignoredRouteFiles: ["**/.*"],
   publicPath: "/_static/build/",
   server: "./server.js",
   serverBuildPath: "server/index.js",
   // appDirectory: "app",
   // assetsBuildDirectory: "public/build",
+  future: {
+    v2_meta: true,
+    v2_routeConvention: true,
+    v2_errorBoundary: true,
+  },
 };