+++ templates/cloudflare-pages/remix.config.js
@@ -1,9 +1,6 @@
 /** @type {import('@remix-run/dev').AppConfig} */
 module.exports = {
   devServerBroadcastDelay: 1000,
-  future: {
-    v2_routeConvention: true,
-  },
   ignoredRouteFiles: ["**/.*"],
   server: "./server.js",
   serverBuildPath: "functions/[[path]].js",
@@ -16,4 +13,9 @@ module.exports = {
   // appDirectory: "app",
   // assetsBuildDirectory: "public/build",
   // publicPath: "/build/",
+  future: {
+    v2_meta: true,
+    v2_routeConvention: true,
+    v2_errorBoundary: true,
+  },
 };