+++ templates/deno/app/routes/_index.tsx
@@ -1,4 +1,9 @@
 import * as React from "react";
+import type { V2_MetaFunction } from "@remix-run/node";
+
+export const meta: V2_MetaFunction = () => {
+  return [{ title: "New Remix App" }];
+};
 
 export default function Index() {
   return (